- testing with karma - restructure and clean code - npm test with karma
