+++ modbustcp/modbustcp-read.html
@@ -1,0 +1,116 @@
+<!--
+ *
+ * Original work Copyright 2015 Valmet Automation Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Modified work Copyright 2016 Argonne National Laboratory.
+ *
+ * Licensed under the the BSD 3-Clause License (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://opensource.org/licenses/BSD-3-Clause
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+-->
+
+<script type="text/javascript">
+    RED.nodes.registerType('modbustcp-read', {
+        category: 'modbus',
+        color: '#E9967A',
+        defaults: {
+            name: {value: ""},
+            dataType: {value: "", required: true},
+            adr: {value: "", required: true, validate: RED.validators.number()},
+            quantity: {value: "", required: true, validate: RED.validators.number()},
+            rate: {
+                value: "", required: true, validate: function (v) {
+                    return v > 0
+                }
+            },
+            rateUnit: {
+                value: "", required: true, validate: function (v) {
+                    return "ms" == v || "s" == v || "m" == v || "h" == v
+                }
+            },
+            server: {value: "", type: "modbustcp-server"}
+        },
+        inputs: 0,
+        outputs: 1,
+        icon: "modbus-icon.png",
+        paletteLabel: "modbustcp",
+        label: function () {
+            return (this.name || "modbustcp");
+        }
+    });
+</script>
+
+
+<script type="text/x-red" data-template-name="modbustcp-read">
+    <div class="form-row">
+        <label for="node-input-name"><i class="icon-tag"></i> Name</label>
+        <input type="text" id="node-input-name" placeholder="Name">
+    </div>
+    <div class="form-row">
+        <label for="node-input-dataType"><i class="icon-list"></i> FC</label>
+        <select id="node-input-dataType">
+            <option value="Coil">FC 1: Read Coils</option>
+            <option value="Input">FC 2: Read Discrete Inputs</option>
+            <option value="HoldingRegister">FC 3: Read Holding Registers</option>
+            <option value="InputRegister">FC 4: Read Input Registers</option>
+        </select>
+    </div>
+    <div class="form-row">
+        <label for="node-input-adr"><i class="icon-home"></i> Address</label>
+        <input type="text" id="node-input-adr" placeholder="0:65535">
+    </div>
+    <div class="form-row">
+        <label for="node-input-quantity"><i class="icon-bookmark"></i> Quantity</label>
+        <input type="text" id="node-input-quantity" placeholder="1-65535">
+    </div>
+    <div class="form-row">
+        <label for="node-input-rate"><i class="icon-bookmark"></i> Poll Rate </label>
+        <input type="number" id="node-input-rate" placeholder="1-65535" style="max-width:120px">
+        <select id="node-input-rateUnit" style="max-width:160px">
+            <option value="ms">millisecond(s)</option>
+            <option value="s">second(s)</option>
+            <option value="m">minute(s)</option>
+            <option value="h">hour(s)</option>
+        </select>
+    </div>
+        <div class="form-row">
+        <label for="node-input-server"><i class="icon-globe"></i> Server</label>
+        <input type="text" id="node-input-server">
+    </div>
+
+</script>
+
+
+<script type="text/x-red" data-help-name="modbustcp-read">
+    <p>Modbus TCP client node.  Connects to a Modbus TCP server to read register values at a given poll rate.
+    <h4>Function codes currently supported include:</h4>
+        <ul>
+          <li>FC 1: Read Coils</li>
+          <li>FC 2: Read Discrete Inputs</li>
+          <li>FC 3: Read Holding Registers</li>
+          <li>FC 4: Read Input Registers</li>
+        </ul>
+     Choose a function code (FC) from the dropdown menu, select the coil/input/register start address (0:65535), and the quantity of coils/inputs/registers to be read from the start address.  Input a poll rate (greater than zero) in seconds to setup the poll rate.  Choose or edit the Modbus TCP server configuration.
+     <br><br>
+
+</script>
