+++ modbustcp/modbustcp-read.js
@@ -33,215 +33,6 @@ module.exports = function (RED) {
     var modbus = require('jsmodbus');
     var util = require('util');
 
-    // ########################################## SERVER #############################################################
-
-    function ModbusTCPServerNode(config) {
-
-        RED.nodes.createNode(this, config);
-
-        this.host = config.host;
-        this.port = config.port;
-        this.unit_id = config.unit_id;
-        this.modbusconn = null;
-
-        var node = this;
-
-        var serverInfo = ' at ' + node.host + ':' + node.port + ' unit_id: ' + node.unit_id;
-
-        function verbose_warn(logMessage) {
-            if (RED.settings.verbose) {
-                node.warn('Server -> ' + logMessage + ' at ' + serverInfo);
-            }
-        }
-
-        function verbose_log(logMessage) {
-            if (RED.settings.verbose) {
-                node.log('Server -> ' + logMessage + ' at ' + serverInfo);
-            }
-        }
-
-        node.initializeModbusTCPConnection = function (handler) {
-
-
-            if (node.modbusconn && node.modbusconn.isConnected()) {
-                verbose_log('Connected to modbustcp slave');
-            }
-            else {
-                verbose_log('Connecting to modbustcp slave');
-
-                node.modbusconn = null;
-                node.modbusconn = modbus.createTCPClient(config.port, config.host, Number(config.unit_id),
-                    function (err) {
-                        if (err) {
-                            node.error('ModbusTCPConnection: ' + util.inspect(err, false, null));
-                            return null;
-                        }
-                        verbose_log('Server connected');
-                        verbose_warn("Server connected");
-                    });
-            }
-
-            handler(node.modbusconn);
-        };
-
-        node.on("close", function () {
-            verbose_warn("Server close");
-            verbose_log('disconnecting from modbustcp slave');
-
-            if (node.modbusconn && node.modbusconn.isConnected()) {
-                node.modbusconn.close();
-                node.modbusconn = null;
-                verbose_log("Server closed");
-                verbose_warn("Server closed");
-            }
-            else {
-                node.modbusconn = null;
-                verbose_log("Server closed");
-                verbose_warn("Server closed");
-            }
-        });
-    }
-
-    RED.nodes.registerType("modbustcp-server", ModbusTCPServerNode);
-
-    // ########################################## WRITE #############################################################
-
-    function ModbusTCPWrite(config) {
-
-        RED.nodes.createNode(this, config);
-
-        this.name = config.name;
-        this.dataType = config.dataType;
-        this.adr = Number(config.adr);
-        this.quantity = config.quantity;
-
-        var node = this;
-        var modbusTCPServer = RED.nodes.getNode(config.server);
-
-        function verbose_warn(logMessage) {
-            if (RED.settings.verbose) {
-                node.warn((node.name) ? node.name + ': ' + logMessage : 'ModbusTCPWrite: ' + logMessage);
-            }
-        }
-
-        function verbose_log(logMessage) {
-            if (RED.settings.verbose) {
-                node.log(logMessage);
-            }
-        }
-
-        modbusTCPServer.initializeModbusTCPConnection(function (connection) {
-
-            node.connection = connection;
-
-            node.receiveEventCloseWrite = function () {
-                if (!node.connection.isConnected()) {
-                    verbose_log('for writing is not connected');
-                    set_unconnected_waiting();
-                }
-            };
-
-            node.connection.on('close', node.receiveEventCloseWrite);
-
-        });
-
-        function set_connected_written(resp) {
-            node.status({fill: "green", shape: "dot", text: util.inspect(resp, false, null)});
-        }
-
-        function set_unconnected_waiting() {
-            node.status({fill: "blue", shape: "dot", text: "waiting ..."});
-        }
-
-        function set_modbus_error(err) {
-            if (err) {
-                node.status({fill: "red", shape: "dot", text: "Error"});
-                verbose_log(err);
-                node.error('ModbusTCPClient: ' + JSON.stringify(err));
-                return false;
-            }
-            return true;
-        }
-
-        this.on("input", function (msg) {
-
-                if (!(msg && msg.hasOwnProperty('payload'))) return;
-
-                if (msg.payload == null) {
-                    node.error('ModbusTCPWrite: Invalid msg.payload!');
-                    return;
-                }
-
-                node.status(null);
-
-                if (!node.connection.isConnected()) {
-                    set_unconnected_waiting();
-                    return;
-                }
-
-                switch (node.dataType) {
-                    case "Coil": //FC: 5
-
-                        if (msg.payload.length < node.quantity) {
-                            node.error("Quantity should be less or equal to coil payload array Addr: ".join(node.adr, " Q: ", node.quantity));
-                        }
-                        if (node.quantity > 1) {
-                            for (i = node.adr; i < node.quantity; i++) {
-
-                                node.connection.writeSingleCoil(i, msg.payload[i], function (resp, err) {
-                                    if (set_modbus_error(err) && resp) {
-                                        set_connected_written(resp);
-                                    }
-                                });
-                            }
-                        } else {
-                            node.connection.writeSingleCoil(node.adr, msg.payload, function (resp, err) {
-                                if (set_modbus_error(err) && resp) {
-                                    set_connected_written(resp);
-                                }
-                            });
-                        }
-                        break;
-
-                    case "HoldingRegister": //FC: 6
-
-                        if (msg.payload.length < node.quantity) {
-                            node.error("Quantity should be less or equal to register payload array Addr: ".join(node.adr, " Q: ", node.quantity));
-                        }
-                        if (node.quantity > 1) {
-                            for (i = node.adr; i < node.quantity; i++) {
-                                node.connection.writeSingleRegister(i, Number(msg.payload[i]), function (resp, err) {
-                                    if (set_modbus_error(err) && resp) {
-                                        set_connected_written(resp);
-                                    }
-                                });
-                            }
-                        } else {
-                            node.connection.writeSingleRegister(node.adr, Number(msg.payload), function (resp, err) {
-                                if (set_modbus_error(err) && resp) {
-                                    set_connected_written(resp);
-                                }
-                            });
-                        }
-                        break;
-
-                    default:
-                        break;
-                }
-            }
-        );
-
-        node.on("close", function () {
-            verbose_warn("write close");
-            node.status({fill: "grey", shape: "dot", text: "Disconnected"});
-        });
-    }
-
-    RED.nodes.registerType("modbustcp-write", ModbusTCPWrite);
-
-
-    // ########################################## READ #############################################################
-
     function ModbusTCPRead(config) {
 
         RED.nodes.createNode(this, config);
@@ -269,7 +60,7 @@ module.exports = function (RED) {
                 node.log(logMessage);
             }
         }
-        
+
         modbusTCPServer.initializeModbusTCPConnection(function (connection) {
 
             node.connection = connection;
@@ -415,5 +206,4 @@ module.exports = function (RED) {
     }
 
     RED.nodes.registerType("modbustcp-read", ModbusTCPRead);
-
 };
