+++ modbustcp/modbustcp-server.html
@@ -1,0 +1,68 @@
+<!--
+ *
+ * Original work Copyright 2015 Valmet Automation Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Modified work Copyright 2016 Argonne National Laboratory.
+ *
+ * Licensed under the the BSD 3-Clause License (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://opensource.org/licenses/BSD-3-Clause
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+-->
+
+
+<script type="text/javascript">
+    RED.nodes.registerType('modbustcp-server', {
+        category: 'config',
+        defaults: {
+            host: {value: "127.0.0.1", required: true},
+            port: {value: 502, required: true, validate: RED.validators.number()},
+            unit_id: {value: 1, required: true, validate: RED.validators.number()}
+        },
+        label: function () {
+            return "modbustcp@" + this.host + ":" + this.port;
+        }
+    });
+</script>
+
+
+<script type="text/x-red" data-template-name="modbustcp-server">
+    <div class="form-row">
+        <label for="node-config-input-host"><i class="icon-bookmark"></i> Host</label>
+        <input type="text" id="node-config-input-host">
+    </div>
+    <div class="form-row">
+        <label for="node-config-input-port"><i class="icon-bookmark"></i> Port</label>
+        <input type="text" id="node-config-input-port">
+    </div>
+    <div class="form-row">
+        <label for="node-config-input-unit_id"><i class="icon-bookmark"></i> Unit Id</label>
+        <input type="text" id="node-config-input-unit_id">
+    </div>
+
+</script>
+
+
+<script type="text/x-red" data-help-name="modbustcp-server">
+    <p>Uses MODBUS TCP to read/write ethernet host:port and register/coil addresses.</p>
+
+</script>
