+++ modbustcp/modbustcp-write.html
@@ -1,0 +1,95 @@
+<!--
+ *
+ * Original work Copyright 2015 Valmet Automation Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Modified work Copyright 2016 Argonne National Laboratory.
+ *
+ * Licensed under the the BSD 3-Clause License (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://opensource.org/licenses/BSD-3-Clause
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+-->
+
+<script type="text/javascript">
+    RED.nodes.registerType('modbustcp-write', {
+        category: 'modbus',
+        color: '#E9967A',
+        defaults: {
+            name: {value: ""},
+            dataType: {value: "", required: true},
+            adr: {value: "", required: true, validate: RED.validators.number()},
+            quantity: {value: "", required: true, validate: RED.validators.number()},
+            server: {value: "", type: "modbustcp-server"}
+        },
+        inputs: 1,
+        outputs: 0,
+        icon: "modbus-icon.png",
+        align: 'right',
+        paletteLabel: "modbustcp",
+        label: function () {
+            return (this.name || "ModbusTCP Write");
+        }
+    });
+</script>
+
+
+<script type="text/x-red" data-template-name="modbustcp-write">
+    <div class="form-row">
+        <label for="node-input-name"><i class="icon-tag"></i> Name</label>
+        <input type="text" id="node-input-name" placeholder="Name">
+    </div>
+    <div class="form-row">
+        <label for="node-input-dataType"><i class="icon-list"></i> Type</label>
+        <select id="node-input-dataType">
+            <option value="Coil">FC 5: Write Single Coil</option>
+            <option value="HoldingRegister">FC 6: Write Single Holding Register</option>
+        </select>
+    </div>
+    <div class="form-row">
+        <label for="node-input-adr"><i class="icon-bookmark"></i> Address</label>
+        <input type="text" id="node-input-adr">
+    </div>
+    <div class="form-row">
+        <label for="node-input-quantity"><i class="icon-bookmark"></i> Quantity</label>
+        <input type="text" id="node-input-quantity" placeholder="1-65535">
+    </div>
+    <div class="form-row">
+        <label for="node-input-server"><i class="icon-bookmark"></i> Server</label>
+        <input type="text" id="node-input-server">
+    </div>
+
+</script>
+
+
+<script type="text/x-red" data-help-name="modbustcp-write">
+    <p>Modbus TCP client node.  Connects to a Modbus TCP server to write <b>msg.payload</b> to a coil or register.
+    <h4>Function codes currently supported include:</h4>
+        <ul>
+          <li>FC 5: Write Single Coil</li>
+          <li>FC 6: Write Single Holding Register</li>
+        </ul>
+     Choose a function code (FC) from the dropdown menu, select the coil/register start address (0:65535).  Choose or edit the Modbus TCP server configuration.
+     <br><br>
+    For FC 5, <b>msg.payload</b> must be a number or string value of 0 or 1.  For FC 6, <b>msg.payload</b> must be a number or string value between 0:65535.
+     </p>
+
+</script>
